+++ src/convert/getopt_test.go
@@ -71,50 +71,3 @@ func TestGetoptLong_Errors(t *testing.T) {
 		})
 	}
 }
-
-func TestGetoptLong_VersionScopedOptions(t *testing.T) {
-	v13 := getoptLong([]string{"-Qr"}, testTable(t, "0.13"), false)
-	if v13[0].Err != "unknown" {
-		t.Fatalf("expected -Q to be unknown in 0.13, got %+v", v13[0])
-	}
-	v17 := getoptLong([]string{"-Qr"}, testTable(t, "0.17"), false)
-	if v17[0].Key != "fake_tls_mod" {
-		t.Fatalf("expected -Q to resolve in 0.17, got %+v", v17[0])
-	}
-
-	n13 := getoptLong([]string{"-n", "example.com"}, testTable(t, "0.13"), false)
-	if n13[0].Key != "tls_sni" {
-		t.Fatalf("expected -n to be tls_sni in 0.13, got %q", n13[0].Key)
-	}
-	n17 := getoptLong([]string{"-n", "example.com"}, testTable(t, "0.17"), false)
-	if n17[0].Key != "fake_sni" {
-		t.Fatalf("expected -n to be fake_sni in 0.17, got %q", n17[0].Key)
-	}
-}
-
-func TestDetectVersion_Markers(t *testing.T) {
-	all, err := loadSpecs()
-	if err != nil {
-		t.Fatal(err)
-	}
-	spec := all["byedpi"]
-	tests := []struct {
-		name     string
-		argv     []string
-		want     string
-		detected bool
-	}{
-		{"fakeTLSMod", []string{"-Qr"}, "0.17", true},
-		{"ipOpt", []string{"-k"}, "0.13", true},
-		{"posRepeats", []string{"-d1:11+sm"}, "0.17", true},
-		{"ambiguousFallsBackToDefault", []string{"-s1", "-f-1"}, "0.17", false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, detected := detectVersion(spec, tt.argv)
-			if got != tt.want || detected != tt.detected {
-				t.Fatalf("got (%s, %v), want (%s, %v)", got, detected, tt.want, tt.detected)
-			}
-		})
-	}
-}
